+++ src/test/java/io/github/astrapi69/checksum/ByteArrayChecksumExtensionsTest.java
@@ -29,9 +29,9 @@ import org.meanbean.test.BeanTester;
 import org.testng.annotations.Test;
 
 import io.github.astrapi69.AbstractTestCase;
-import io.github.astrapi69.crypto.algorithm.Algorithm;
-import io.github.astrapi69.crypto.algorithm.HashAlgorithm;
-import io.github.astrapi69.crypto.algorithm.MdAlgorithm;
+import io.github.astrapi69.crypt.api.algorithm.Algorithm;
+import io.github.astrapi69.crypt.api.algorithm.HashAlgorithm;
+import io.github.astrapi69.crypt.api.algorithm.MdAlgorithm;
 
 /**
  * The unit test class for the class {@link ByteArrayChecksumExtensions}
@@ -80,14 +80,14 @@ public class ByteArrayChecksumExtensionsTest extends AbstractTestCase<Long, Long
 	}
 
 	/**
-	 * Test method for {@link ByteArrayChecksumExtensions#getChecksum(Algorithm, byte[]...)}
+	 * Test method for {@link ByteArrayChecksumExtensions#getChecksums(Algorithm, byte[]...)}
 	 *
 	 * @throws NoSuchAlgorithmException
 	 *             Is thrown if the algorithm is not supported or does not exists.
 	 *             {@link java.security.MessageDigest} object.
 	 */
 	@Test
-	public void testGetChecksumAlgorithmByteArrays() throws NoSuchAlgorithmException
+	public void testGetChecksumsAlgorithmByteArrays() throws NoSuchAlgorithmException
 	{
 		String expected;
 		String actual;
@@ -98,7 +98,7 @@ public class ByteArrayChecksumExtensionsTest extends AbstractTestCase<Long, Long
 		final byte[] secretMessageBytes = secretMessage.getBytes(StandardCharsets.UTF_8);
 		final byte[] secondMessageBytes = secondMessage.getBytes(StandardCharsets.UTF_8);
 		expected = "5cc16e663491726545c13ec2012f4601d11f8ce29210b4b50c5e67533b699d02";
-		actual = ByteArrayChecksumExtensions.getChecksum(MdAlgorithm.MD2, secretMessageBytes,
+		actual = ByteArrayChecksumExtensions.getChecksums(MdAlgorithm.MD2, secretMessageBytes,
 			secondMessageBytes);
 		assertEquals(expected, actual);
 		actualLength = actual.length();
@@ -106,14 +106,14 @@ public class ByteArrayChecksumExtensionsTest extends AbstractTestCase<Long, Long
 		assertEquals(expectedLength, actualLength);
 
 		expected = "25659bd9db98ecc3c2077d44e69607b8acbd18db4cc2f85cedef654fccc4a4d8";
-		actual = ByteArrayChecksumExtensions.getChecksum(MdAlgorithm.MD5, secretMessageBytes,
+		actual = ByteArrayChecksumExtensions.getChecksums(MdAlgorithm.MD5, secretMessageBytes,
 			secondMessageBytes);
 		assertEquals(expected, actual);
 		actualLength = actual.length();
 		assertEquals(expectedLength, actualLength);
 
 		expected = "874026e54b67d4f9aaf87cb14a683fb51de6f9cb0beec7b5ea3f0fdbc95d0dd47f3c5bc275da8a33";
-		actual = ByteArrayChecksumExtensions.getChecksum(HashAlgorithm.SHA_1, secretMessageBytes,
+		actual = ByteArrayChecksumExtensions.getChecksums(HashAlgorithm.SHA_1, secretMessageBytes,
 			secondMessageBytes);
 		assertEquals(expected, actual);
 		actualLength = actual.length();
@@ -121,7 +121,7 @@ public class ByteArrayChecksumExtensionsTest extends AbstractTestCase<Long, Long
 		assertEquals(expectedLength, actualLength);
 
 		expected = "8a3b3c92a8b0eb00da917c23201a9407ef7963373464076aec4c54c066e8b7aa2c26b46b68ffc68ff99b453c1d30413413422d706483bfa0f98a5e886266e7ae";
-		actual = ByteArrayChecksumExtensions.getChecksum(HashAlgorithm.SHA_256, secretMessageBytes,
+		actual = ByteArrayChecksumExtensions.getChecksums(HashAlgorithm.SHA_256, secretMessageBytes,
 			secondMessageBytes);
 		assertEquals(expected, actual);
 		actualLength = actual.length();
@@ -129,7 +129,7 @@ public class ByteArrayChecksumExtensionsTest extends AbstractTestCase<Long, Long
 		assertEquals(expectedLength, actualLength);
 
 		expected = "b58a362687ab42b9bf0d8af0b4860ed262d1fd128e16ab0082723e7785a862cd129b03577312452cc24aecdb36d5406d98c11ffdfdd540676b1a137cb1a22b2a70350c9a44171d6b1180c6be5cbb2ee3f79d532c8a1dd9ef2e8e08e752a3babb";
-		actual = ByteArrayChecksumExtensions.getChecksum(HashAlgorithm.SHA_384, secretMessageBytes,
+		actual = ByteArrayChecksumExtensions.getChecksums(HashAlgorithm.SHA_384, secretMessageBytes,
 			secondMessageBytes);
 		assertEquals(expected, actual);
 		actualLength = actual.length();
@@ -137,7 +137,7 @@ public class ByteArrayChecksumExtensionsTest extends AbstractTestCase<Long, Long
 		assertEquals(expectedLength, actualLength);
 
 		expected = "ab29b34a26547ca4ce517d776885a5642929d9ed571a990fc764f7d0b854d6546276ca9aa45b3d88db3dc3dbf3c2f2152017d3e3e054ed6cd7a38a1f7925a746f7fbba6e0636f890e56fbbf3283e524c6fa3204ae298382d624741d0dc6638326e282c41be5e4254d8820772c5518a2c5a8c0c7f7eda19594a7eb539453e1ed7";
-		actual = ByteArrayChecksumExtensions.getChecksum(HashAlgorithm.SHA_512, secretMessageBytes,
+		actual = ByteArrayChecksumExtensions.getChecksums(HashAlgorithm.SHA_512, secretMessageBytes,
 			secondMessageBytes);
 		assertEquals(expected, actual);
 		actualLength = actual.length();